First worked version without errors

+++ ParaCL/Interpreter.hh
@@ -1,4 +1,4 @@
 #include"ParaCL.hpp"
 
 int Interpreter(yy::Driver *  driver, ISyntaxTreeNode * node);
-int CalculateRes(yy::Driver * driver, ISyntaxTreeNode * node);
+double CalculateRes(yy::Driver * driver, ISyntaxTreeNode * node);
